ml-models/training/scripts/train_toxicity.py
"""
Training script for toxicity detection model with fairness constraints
"""

import torch
from torch.utils.data import DataLoader
from transformers import AutoTokenizer, AutoModelForSequenceClassification, Trainer, TrainingArguments
import yaml
import argparse
from pathlib import Path

class ToxicityTrainer:
    def __init__(self, config_path):
        with open(config_path, 'r') as f:
            self.config = yaml.safe_load(f)
        
        self.model_name = self.config['model']['name']
        self.output_dir = self.config['training']['output_dir']
        
    def load_model(self):
        """Load pre-trained model and tokenizer"""
        self.tokenizer = AutoTokenizer.from_pretrained(self.model_name)
        self.model = AutoModelForSequenceClassification.from_pretrained(
            self.model_name,
            num_labels=self.config['model']['num_labels']
        )
        
    def load_dataset(self):
        """Load and preprocess dataset"""
        # TODO: Implement dataset loading
        # Should include balanced samples across demographics
        pass
    
    def train(self):
        """Train the model with fairness constraints"""
        training_args = TrainingArguments(
            output_dir=self.output_dir,
            num_train_epochs=self.config['training']['epochs'],
            per_device_train_batch_size=self.config['training']['batch_size'],
            learning_rate=self.config['training']['learning_rate'],
            logging_dir=f"{self.output_dir}/logs",
            logging_steps=100,
            evaluation_strategy="epoch",
            save_strategy="epoch",
            load_best_model_at_end=True,
        )
        
        # TODO: Implement training loop
        # Include fairness metrics evaluation
        # Apply bias mitigation techniques
        
    def evaluate_fairness(self):
        """Evaluate model for fairness across demographics"""
        # TODO: Implement fairness evaluation
        # Calculate demographic parity
        # Measure equal opportunity
        pass
    
    def save_model(self):
        """Save trained model and metadata"""
        self.model.save_pretrained(self.output_dir)
        self.tokenizer.save_pretrained(self.output_dir)
        
        # Save model metadata
        metadata = {
            'model_name': self.model_name,
            'version': self.config['model']['version'],
            'trained_date': str(Path.ctime(Path(self.output_dir))),
            'metrics': {},  # Add evaluation metrics
            'fairness_metrics': {}  # Add fairness metrics
        }
        
        with open(f"{self.output_dir}/metadata.yaml", 'w') as f:
            yaml.dump(metadata, f)

if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument('--config', type=str, required=True, help='Path to config file')
    args = parser.parse_args()
    
    trainer = ToxicityTrainer(args.config)
    trainer.load_model()
    trainer.load_dataset()
    trainer.train()
    trainer.evaluate_fairness()
    trainer.save_model()
