ml-models/training/requirements.txt
# ML Training Requirements

# Deep Learning Frameworks
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
datasets==2.15.0
accelerate==0.25.0

# Training Utilities
tensorboard==2.15.1
wandb==0.16.1
pytorch-lightning==2.1.2

# NLP
tokenizers==0.15.0
sentencepiece==0.1.99
nltk==3.8.1
spacy==3.7.2

# Fairness & Bias
fairlearn==0.10.0
aif360==0.5.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Configuration
pyyaml==6.0.1
hydra-core==1.3.2

# Utilities
tqdm==4.66.1
python-dotenv==1.0.0
